requirements.txt
openai
pandas
matplotlib
seaborn
python-dotenv
yaml
tqdm
huggingface-hub
torch